refactored a few files
