+++ Behavioral/ChainOfResponsibility/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Chain of Responsibility Pattern</title>
     <link rel="stylesheet" href="style.css" />
-    <script src="./dist/chainOfResponsibilityPattern.js" defer></script>
+    <script src="./index.js" defer></script>
   </head>
   <body>
     <div class="window">
@@ -14,7 +14,7 @@
         <div class="form-title">
           <h4>Login</h4>
         </div>
-        <form id="validationForm">
+        <form id="validationForm" class="validationForm">
           <div class="form-input">
             <label for="username">Username:</label>
             <input type="text" id="username" placeholder="Username" />
