+++ Behavioral/ChainOfResponsibility/src/models/BaseHandler.ts
@@ -1,0 +1,14 @@
+import ValidationHandler from "./ValidationHandler";
+import ValidationInput from "./ValidationInput";
+
+abstract class BaseHandler implements ValidationHandler {
+  protected nextHandler: ValidationHandler;
+
+  setNextHandler(handler: ValidationHandler): void {
+    this.nextHandler = handler;
+  }
+
+  abstract handleValidation(inputValues: ValidationInput): boolean;
+}
+
+export default BaseHandler;
