+++ Behavioral/ChainOfResponsibility/src/models/ValidationHandler.ts
@@ -1,0 +1,8 @@
+import ValidationInput from "./ValidationInput";
+
+interface ValidationHandler {
+  setNextHandler(handler: ValidationHandler): void;
+  handleValidation(inputValue: ValidationInput): boolean;
+}
+
+export default ValidationHandler;
