+++ Behavioral/ChainOfResponsibility/src/models/ValidationInput.ts
@@ -1,0 +1,7 @@
+interface ValidationInput {
+  username: string;
+  email: string;
+  password: string;
+}
+
+export default ValidationInput;
